Update to version 2.1.0-alpha. Improved path alias usage and override handling. Refactored imports to use specific path aliases instead of relative traversal
