+++ tests/cart.spec.ts
@@ -1,13 +1,13 @@
 // @ts-check
 
 import { test, expect } from '@playwright/test';
-import { UIReference, slugs, outcomeMarker } from 'config';
+import { UIReference, slugs, outcomeMarker } from '@config';
 
-import CartPage from './poms/frontend/cart.page';
-import LoginPage from './poms/frontend/login.page';
-import ProductPage from './poms/frontend/product.page';
-import { requireEnv } from './utils/env.utils';
-import NotificationValidator from './utils/notification.validator';
+import CartPage from '@poms/frontend/cart.page';
+import LoginPage from '@poms/frontend/login.page';
+import ProductPage from '@poms/frontend/product.page';
+import { requireEnv } from '@utils/env.utils';
+import NotificationValidator from '@utils/notification.validator';
 
 test.describe('Cart functionalities (guest)', () => {
   /**
