+++ tests/compare.spec.ts
@@ -1,12 +1,12 @@
 // @ts-check
 
 import { test, expect } from '@playwright/test';
-import { UIReference, outcomeMarker, slugs } from 'config';
+import { UIReference, outcomeMarker, slugs } from '@config';
 
-import ComparePage from './poms/frontend/compare.page';
-import LoginPage from './poms/frontend/login.page';
-import ProductPage from './poms/frontend/product.page';
-import { requireEnv } from './utils/env.utils';
+import ComparePage from '@poms/frontend/compare.page';
+import LoginPage from '@poms/frontend/login.page';
+import ProductPage from '@poms/frontend/product.page';
+import { requireEnv } from '@utils/env.utils';
 
 // TODO: Create a fixture for this
 test.beforeEach('Add 2 products to compare, then navigate to comparison page', async ({ page }) => {
