+++ tests/config/index.ts
@@ -1,0 +1,7 @@
+// @ts-check
+
+export { default as UIReference } from './element-identifiers.json';
+export { default as outcomeMarker } from './outcome-markers.json';
+export { default as inputValues } from './input-values.json';
+export { default as slugs } from './slugs.json';
+export { default as toggles } from './test-toggles.json';
