+++ tests/home.spec.ts
@@ -1,10 +1,10 @@
 // @ts-check
 
 import { test, expect } from '@playwright/test';
-import { outcomeMarker } from 'config';
+import { outcomeMarker } from '@config';
 
-import MainMenuPage from './poms/frontend/mainmenu.page';
-import HomePage from './poms/frontend/home.page';
+import MainMenuPage from '@poms/frontend/mainmenu.page';
+import HomePage from '@poms/frontend/home.page';
 
 test('Add_product_on_homepage_to_cart',{ tag: ['@homepage', '@cold']}, async ({page}) => {
   const homepage = new HomePage(page);
