+++ tests/login.spec.ts
@@ -1,10 +1,10 @@
 // @ts-check
 
 import { test as base, expect } from '@playwright/test';
-import { outcomeMarker, inputValues } from 'config';
-import { requireEnv } from './utils/env.utils';
+import { outcomeMarker, inputValues } from '@config';
+import { requireEnv } from '@utils/env.utils';
 
-import LoginPage from './poms/frontend/login.page';
+import LoginPage from '@poms/frontend/login.page';
 
 base('User_logs_in_with_valid_credentials', {tag: '@hot'}, async ({page, browserName}) => {
   const browserEngine = browserName?.toUpperCase() || "UNKNOWN";
