+++ tests/minicart.spec.ts
@@ -1,11 +1,11 @@
 // @ts-check
 
 import {test, expect} from '@playwright/test';
-import {UIReference, outcomeMarker, slugs} from 'config';
+import {UIReference, outcomeMarker, slugs} from '@config';
 
-import MainMenuPage from './poms/frontend/mainmenu.page';
-import ProductPage from './poms/frontend/product.page';
-import MiniCartPage from './poms/frontend/minicart.page';
+import MainMenuPage from '@poms/frontend/mainmenu.page';
+import ProductPage from '@poms/frontend/product.page';
+import MiniCartPage from '@poms/frontend/minicart.page';
 
 test.describe('Minicart Actions', {annotation: {type: 'Minicart', description: 'Minicart simple product tests'},}, () => {
