+++ tests/poms/adminhtml/magentoAdmin.page.ts
@@ -1,7 +1,7 @@
 // @ts-check
 
 import { expect, type Locator, type Page } from '@playwright/test';
-import { UIReference, inputValues } from 'config';
+import { UIReference, inputValues } from '@config';
 
 class MagentoAdminPage {
   readonly page: Page;
