+++ tests/poms/frontend/account.page.ts
@@ -2,8 +2,9 @@
 
 import { expect, type Locator, type Page } from '@playwright/test';
 import { faker } from '@faker-js/faker';
-import { UIReference, outcomeMarker, inputValues } from 'config';
-import LoginPage from './login.page';
+import { UIReference, outcomeMarker, inputValues } from '@config';
+
+import LoginPage from '@poms/frontend/login.page';
 
 class AccountPage {
   readonly page: Page;
