+++ tests/poms/frontend/cart.page.ts
@@ -1,7 +1,7 @@
 // @ts-check
 
 import { expect, type Locator, type Page } from '@playwright/test';
-import { UIReference, outcomeMarker } from 'config';
+import { UIReference, outcomeMarker } from '@config';
 
 class CartPage {
   readonly page: Page;
