+++ tests/poms/frontend/compare.page.ts
@@ -1,7 +1,7 @@
 // @ts-check
 
 import { expect, type Page } from '@playwright/test';
-import { UIReference, outcomeMarker } from 'config';
+import { UIReference, outcomeMarker } from '@config';
 
 class ComparePage {
   page: Page;
