+++ tests/poms/frontend/home.page.ts
@@ -1,7 +1,7 @@
 // @ts-check
 
 import { type Page } from '@playwright/test';
-import { UIReference } from 'config';
+import { UIReference } from '@config';
 
 class HomePage {
