+++ tests/poms/frontend/register.page.ts
@@ -1,7 +1,7 @@
 // @ts-check
 
 import { expect, type Locator, type Page } from '@playwright/test';
-import { UIReference, outcomeMarker, slugs} from 'config';
+import { UIReference, outcomeMarker, slugs} from '@config';
 
 class RegisterPage {
   readonly page: Page;
