+++ tests/poms/frontend/search.page.ts
@@ -1,7 +1,7 @@
 // @ts-check
 
 import { expect, type Locator, type Page } from '@playwright/test';
-import { UIReference } from 'config';
+import { UIReference } from '@config';
 
 class SearchPage {
   readonly page: Page;
