+++ tests/register.spec.ts
@@ -2,10 +2,10 @@
 
 import {test} from '@playwright/test';
 import {faker} from '@faker-js/faker';
-import { inputValues } from 'config';
+import { inputValues } from '@config';
 
-import RegisterPage from './poms/frontend/register.page';
-import { requireEnv } from './utils/env.utils';
+import RegisterPage from '@poms/frontend/register.page';
+import { requireEnv } from '@utils/env.utils';
 
 // Reset storageState to ensure we're not logged in before running these tests.
 test.use({ storageState: { cookies: [], origins: [] } });
